CEDS-1211: moved tests around
