+++ build.sbt
@@ -1,3 +1,4 @@
+import sbt.Tests.{Group, SubProcess}
 import sbt._
 import uk.gov.hmrc.DefaultBuildSettings._
 import uk.gov.hmrc.sbtdistributables.SbtDistributablesPlugin.publishingSettings
@@ -5,6 +6,28 @@ import uk.gov.hmrc.{SbtArtifactory, SbtAutoBuildPlugin}
 import uk.gov.hmrc.versioning.SbtGitVersioning
 
 val appName = "customs-declare-exports-movements"
+
+
+lazy val allResolvers = resolvers ++= Seq(
+  Resolver.bintrayRepo("hmrc", "releases"),
+  Resolver.jcenterRepo
+)
+
+lazy val ComponentTest = config("component") extend Test
+lazy val CdsIntegrationTest = config("it") extend Test
+
+val testConfig = Seq(ComponentTest, CdsIntegrationTest, Test)
+
+def forkedJvmPerTestConfig(tests: Seq[TestDefinition], packages: String*): Seq[Group] =
+  tests.groupBy(_.name.takeWhile(_ != '.')).filter(packageAndTests => packages contains packageAndTests._1) map {
+    case (packg, theTests) =>
+      Group(packg, theTests, SubProcess(ForkOptions()))
+  } toSeq
+
+lazy val testAll = TaskKey[Unit]("test-all")
+lazy val allTest = Seq(testAll := (test in ComponentTest)
+  .dependsOn((test in CdsIntegrationTest).dependsOn(test in Test)).value)
+
 
 lazy val microservice = Project(appName, file("."))
   .enablePlugins(PlayScala, SbtAutoBuildPlugin, SbtGitVersioning, SbtDistributablesPlugin, SbtArtifactory)
@@ -16,17 +39,38 @@ lazy val microservice = Project(appName, file("."))
     scalafmtFailTest in ThisBuild := false
   )
   .settings(publishingSettings: _*)
-  .configs(IntegrationTest)
+  .configs(testConfig: _*)
   .settings(inConfig(IntegrationTest)(Defaults.itSettings): _*)
-  .settings(
-    Keys.fork in IntegrationTest := false,
-    unmanagedSourceDirectories in IntegrationTest := (baseDirectory in IntegrationTest)(base => Seq(base / "it")).value,
-    testGrouping in IntegrationTest := TestPhases.oneForkedJvmPerTest((definedTests in IntegrationTest).value),
-    parallelExecution in IntegrationTest := false,
-    addTestReportOption(IntegrationTest, "int-test-reports")
-  )
   .settings(resolvers += Resolver.jcenterRepo)
-  .settings(scoverageSettings)
+  .settings(commonSettings,
+    unitTestSettings,
+    integrationTestSettings,
+    allResolvers,
+    scoverageSettings)
+
+def onPackageName(rootPackage: String): String => Boolean = {
+  testName => testName startsWith rootPackage
+}
+
+lazy val unitTestSettings =
+  inConfig(Test)(Defaults.testTasks) ++
+    Seq(
+      testOptions in Test := Seq(Tests.Filter(onPackageName("unit"))),
+      testOptions in Test += Tests.Argument(TestFrameworks.ScalaTest, "-oD"),
+      unmanagedSourceDirectories in Test := Seq((baseDirectory in Test).value / "test"),
+      addTestReportOption(Test, "test-reports")
+    )
+
+lazy val integrationTestSettings =
+  inConfig(CdsIntegrationTest)(Defaults.testTasks) ++
+    Seq(
+      testOptions in CdsIntegrationTest := Seq(Tests.Filters(Seq(onPackageName("integration"), onPackageName("component")))),
+      testOptions in CdsIntegrationTest += Tests.Argument(TestFrameworks.ScalaTest, "-oD"),
+      fork in CdsIntegrationTest := false,
+      parallelExecution in CdsIntegrationTest := false,
+      addTestReportOption(CdsIntegrationTest, "int-test-reports"),
+      testGrouping in CdsIntegrationTest := forkedJvmPerTestConfig((definedTests in Test).value, "integration", "component")
+    )
 
 lazy val scoverageSettings: Seq[Setting[_]] = Seq(
   coverageExcludedPackages := List(
@@ -42,3 +86,8 @@ lazy val scoverageSettings: Seq[Setting[_]] = Seq(
   coverageHighlighting := true,
   parallelExecution in Test := false
 )
+
+lazy val commonSettings: Seq[Setting[_]] = scalaSettings ++
+  publishingSettings ++
+  defaultSettings() ++
+  gitStampSettings
